running unit tests phase 1
