+++ pkg/executor/client/interface.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 
 	"github.com/kubeshop/testkube/pkg/api/v1/testkube"
+	"github.com/kubeshop/testkube/pkg/executor/options"
 	"github.com/kubeshop/testkube/pkg/executor/output"
 )
 
@@ -21,7 +22,7 @@ type ResultEvent struct {
 type Executor interface {
 	// Execute starts new external test execution, reads data and returns ID
 	// execution is started asynchronously client can check later for results
-	Execute(ctx context.Context, execution *testkube.Execution, options ExecuteOptions) (result *testkube.ExecutionResult, err error)
+	Execute(ctx context.Context, execution *testkube.Execution, options options.ExecuteOptions) (result *testkube.ExecutionResult, err error)
 
 	// Abort aborts pending execution, do nothing when there is no pending execution
 	Abort(ctx context.Context, execution *testkube.Execution) (result *testkube.ExecutionResult, err error)
